PlayCmdの移行開始 & リピートバグ修正(多分
